+++ all-in-one-wp-security/admin/wp-security-list-acct-activity.php
@@ -19,9 +19,13 @@ class AIOWPSecurity_List_Account_Activity extends AIOWPSecurity_List_Table {
         
     function column_user_id($item){
         $tab = strip_tags($_REQUEST['tab']);
+        $delete_url = sprintf('admin.php?page=%s&tab=%s&action=%s&activity_login_rec=%s', AIOWPSEC_USER_LOGIN_MENU_SLUG, $tab, 'delete_acct_activity_rec', $item['id']);
+        //Add nonce to delete URL
+        $delete_url_nonce = wp_nonce_url($delete_url, "delete_acct_activity_log", "aiowps_nonce");
+        
         //Build row actions
         $actions = array(
-            'delete' => sprintf('<a href="admin.php?page=%s&tab=%s&action=%s&activity_login_rec=%s" onclick="return confirm(\'Are you sure you want to delete this item?\')">Delete</a>',AIOWPSEC_USER_LOGIN_MENU_SLUG,$tab,'delete_acct_activity_rec',$item['id']),
+            'delete' => '<a href="'.$delete_url_nonce.'" onclick="return confirm(\'Are you sure you want to delete this item?\')">Delete</a>',
         );
         
         //Return the user_login contents
@@ -93,24 +97,33 @@ class AIOWPSecurity_List_Account_Activity extends AIOWPSecurity_List_Table {
      */
     function delete_login_activity_records($entries)
     {
-        global $wpdb;
+        global $wpdb, $aio_wp_security;
         $login_activity_table = AIOWPSEC_TBL_USER_LOGIN_ACTIVITY;
         if (is_array($entries))
         {
-            //Delete multiple records
-            $id_list = "(" .implode(",",$entries) .")"; //Create comma separate list for DB operation
-            $delete_command = "DELETE FROM ".$login_activity_table." WHERE id IN ".$id_list;
-            $result = $wpdb->query($delete_command);
-            if($result != NULL)
+            if (isset($_REQUEST['_wp_http_referer']))
             {
-                $success_msg = '<div id="message" class="updated fade"><p><strong>';
-                $success_msg .= __('The selected entries were deleted successfully!','aiowpsecurity');
-                $success_msg .= '</strong></p></div>';
-                _e($success_msg);
+                //Delete multiple records
+                $id_list = "(" .implode(",",$entries) .")"; //Create comma separate list for DB operation
+                $delete_command = "DELETE FROM ".$login_activity_table." WHERE id IN ".$id_list;
+                $result = $wpdb->query($delete_command);
+                if($result != NULL)
+                {
+                    $success_msg = '<div id="message" class="updated fade"><p><strong>';
+                    $success_msg .= __('The selected entries were deleted successfully!','aiowpsecurity');
+                    $success_msg .= '</strong></p></div>';
+                    _e($success_msg);
+                }
             }
         } 
         elseif ($entries != NULL)
         {
+            $nonce=isset($_GET['aiowps_nonce'])?$_GET['aiowps_nonce']:'';
+            if (!isset($nonce) ||!wp_verify_nonce($nonce, 'delete_acct_activity_log'))
+            {
+                $aio_wp_security->debug_logger->log_debug("Nonce check failed for delete selected account activity logs operation!",4);
+                die(__('Nonce check failed for delete selected account activity logs operation!','aiowpsecurity'));
+            }
             //Delete single record
             $delete_command = "DELETE FROM ".$login_activity_table." WHERE id = '".absint($entries)."'";
             $result = $wpdb->query($delete_command);
